requirements.txt
playwright
fastapi
uvicorn
python-dotenv
supabase
pandas
pypdf2
requests
